third_party/WebKit/LayoutTests/editing/inserting/break-out-of-nested-lists-expected.txt

Tests breaking out of empty list item in case of nested lists.
| "
"
| <ol>
|   "
"
|   <li>
|     "one"
|   "
"
|   <ul>
|     "
"
|     <li>
|       id="li"
|       "1"
| <div>
|   "foo<#selection-caret>"
| <ol>
|   <ul>
|     "
"
|     <li>
|       "2"
|     "
"
|   "
"
|   <li>
|     "two"
|   "
"
| "
"
